Rejigged trains braking and accelerating because there were several bugs stopping train reversing out of station. Most importantly, changed braking and accelerating to Mathf.SmoothDamp instead of the custom function in Train. Second: it seemed that DoorOpen was called multiple times because it is called in FixedUpdate on a check for velocity close to 0. Fixed this by changing to Idle state before DoorOpen called. Changed signal box so that it is set to accelerate when the train hits it- THIS IS TEMPORARY- the player will control this in future- put this in just so train doesnt hit brakes on the way out of the station.

+++ Assets/Destination.cs
@@ -8,12 +8,12 @@ using UnityEngine;
 public class Destination {
 	public Material materialColor{ get; private set; }
 	public string name{ get; private set; }
-	public float routeLength { get; private set; }
+	public int routeLength { get; private set; }
 
 	int estimatedDailyFootfall;
 	static List<Material> availableMaterialColors = new List<Material>();
 
-	public Destination(string _name, float _routeLength, int _estimatedDailyFootfall) {
+	public Destination(string _name, int _routeLength, int _estimatedDailyFootfall) {
 		name = _name;
 		routeLength = _routeLength;	//in km
 		estimatedDailyFootfall = _estimatedDailyFootfall;
